Modified Pipeline for MultiFrequency UNet
I modified the exisitng pipeline I had created using Transform Images Notebook, to directly use the components created by the FB2WEDT transformation.  This was a much more creative and effecient approach to solve the problem. So everytime an image was called for training, it would generate the components which would then be fed into the modified UNet